+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Renamed asynchronous *ProjectAuditor.Audit* to *AuditAsync*
+
 ## [0.8.0-preview] - 2022-05-20
 * Added *ProjectAuditor.NumCategories* API
 * Added module-specific incremental analysis support